+++ apps/website/src/docs/components/sidebar/sidebar.tsx
@@ -37,7 +37,8 @@ function SidebarHeader({ isOpen, isActive, children, ...props }: any) {
       <chakra.h4
         fontSize="sm"
         fontWeight="bold"
-        my="1rem"
+        mt="1rem"
+        mb="0.5rem"
         letterSpacing="wider"
         color={isActive ? color : 'muted'}
         display="flex"
@@ -121,9 +122,7 @@ function SidebarGroup({
               if (!lvl2.routes) {
                 return (
                   <SidebarLink
-                    mb="1"
-                    ml="4"
-                    color="muted"
+                    mb="2px"
                     key={lvl2.path || index}
                     href={lvl2.path}
                   >
@@ -193,11 +192,7 @@ function SidebarGroup({
                 >
                   {/* <Stack spacing="1"> */}
                   {sortedRoutes.map((lvl3, i) => (
-                    <SidebarLink
-                      key={lvl3.path || i}
-                      href={lvl3.path}
-                      color="muted"
-                    >
+                    <SidebarLink key={lvl3.path || i} href={lvl3.path} mb="2px">
                       <span>{convertBackticksToInlineCode(lvl3.title)}</span>
                       {lvl3.new && (
                         <Badge
@@ -260,7 +255,7 @@ export function SidebarContent(props: SidebarContentProps) {
   const { routes, pathname, contentRef } = props
   const color = useColorModeValue('gray.700', 'inherit')
 
-  const searchRef = React.useRef(null)
+  const searchRef = React.useRef<any>(null)
 
   useHotkeys('/', () => {
     searchRef.current?.focus()