+++ packages/saas-ui-core/src/menu/context-menu.tsx
@@ -2,7 +2,6 @@ import * as React from 'react'
 import { useCallback, useRef, useState } from 'react'
 import {
   chakra,
-  Portal,
   Menu,
   MenuProps,
   MenuList,
@@ -11,19 +10,27 @@ import {
   HTMLChakraProps,
   useMenuContext,
   useEventListener,
+  useOutsideClick,
 } from '@chakra-ui/react'
 
 import { createContext } from '@chakra-ui/react-utils'
-import { runIfFn } from '@chakra-ui/utils'
+import { AnyPointerEvent, callAllHandlers, runIfFn } from '@chakra-ui/utils'
+
+// @todo migrate this to Ark-ui ContextMenu
+import { useLongPress } from '@react-aria/interactions'
+
+import { getEventPoint } from '@zag-js/dom-event'
 
 type Position = [number, number]
+type Anchor = { x: number; y: number }
 
 export interface UseContextMenuReturn {
   isOpen: boolean
-  position: Position
+  anchor: Anchor
   triggerRef: React.RefObject<HTMLSpanElement>
+  menuRef: React.RefObject<HTMLDivElement>
   onClose: () => void
-  onOpen: (event: React.MouseEvent) => void
+  onOpen: (event: AnyPointerEvent) => void
 }
 
 export const [ContextMenuProvider, useContextMenuContext] =
@@ -37,23 +44,42 @@ export interface UseContextMenuProps extends ContextMenuProps {
 }
 
 export const useContextMenu = (props: UseContextMenuProps) => {
+  const { closeOnBlur = true } = props
   const [isOpen, setIsOpen] = useState(false)
-  const [position, setPosition] = useState<Position>([0, 0])
+  const [anchor, setAnchor] = useState<Anchor>({ x: 0, y: 0 })
   const triggerRef = useRef<HTMLSpanElement>(null)
+  const menuRef = useRef<HTMLDivElement>(null)
 
-  // useOutsideClick off menu doesn't catch contextmenu
+  // useOutsideClick of menu doesn't catch contextmenu
   useEventListener('contextmenu', (e) => {
     if (
       !triggerRef.current?.contains(e.target as any) &&
       e.target !== triggerRef.current
     ) {
       setIsOpen(false)
+    } else {
+      e.preventDefault()
+      e.stopPropagation()
     }
   })
 
-  const onOpen = useCallback((event: React.MouseEvent) => {
+  useOutsideClick({
+    enabled: isOpen && closeOnBlur,
+    ref: menuRef,
+    handler: (event) => {
+      if (
+        !triggerRef.current?.contains(event.target as HTMLElement) &&
+        menuRef.current?.parentElement !== event.target
+      ) {
+        onClose()
+      }
+    },
+  })
+
+  const onOpen = useCallback((event: AnyPointerEvent) => {
+    const point = getEventPoint(event)
+    setAnchor(point)
     setIsOpen(true)
-    setPosition([event.pageX, event.pageY])
   }, [])
 
   const onClose = useCallback(() => {
@@ -63,8 +89,9 @@ export const useContextMenu = (props: UseContextMenuProps) => {
 
   return {
     isOpen,
-    position,
+    anchor,
     triggerRef,
+    menuRef,
     onClose,
     onOpen,
   }
@@ -80,7 +107,13 @@ export const ContextMenu: React.FC<ContextMenuProps> = (props) => {
   const { isOpen, onClose } = context
 
   return (
-    <Menu gutter={0} {...rest} isOpen={isOpen} onClose={onClose}>
+    <Menu
+      gutter={0}
+      {...rest}
+      isOpen={isOpen}
+      onClose={onClose}
+      closeOnBlur={false}
+    >
       {(fnProps) => (
         <ContextMenuProvider value={context}>
           {runIfFn(children, fnProps)}
@@ -92,29 +125,86 @@ export const ContextMenu: React.FC<ContextMenuProps> = (props) => {
 
 ContextMenu.displayName = 'ContextMenu'
 
+const generateClientRect = (x = 0, y = 0) => {
+  return () => {
+    return {
+      width: 0,
+      height: 0,
+      top: y,
+      left: x,
+      right: x,
+      bottom: y,
+    }
+  }
+}
+
+const useContextMenuTrigger = (props: ContextMenuTriggerProps) => {
+  const { triggerRef, onOpen, onClose, anchor } = useContextMenuContext()
+
+  const menu = useMenuContext()
+
+  const { popper, openAndFocusFirstItem } = menu
+
+  const { longPressProps } = useLongPress({
+    accessibilityDescription: 'Long press to open context menu',
+    onLongPressStart: (e) => {
+      if (e.pointerType === 'mouse') {
+        onClose()
+      }
+    },
+    onLongPress: (e) => {
+      if (e.pointerType === 'mouse') return
+
+      if (e.type === 'longpress') {
+        onOpen(e as unknown as AnyPointerEvent)
+        openAndFocusFirstItem()
+      }
+    },
+  })
+
+  const anchorRef = React.useRef({
+    getBoundingClientRect: generateClientRect(anchor.x, anchor.y),
+  })
+
+  React.useEffect(() => {
+    popper.referenceRef(anchorRef.current)
+  }, [])
+
+  React.useEffect(() => {
+    anchorRef.current.getBoundingClientRect = generateClientRect(
+      anchor.x,
+      anchor.y
+    )
+    menu.popper.update()
+  }, [anchor])
+
+  return {
+    triggerProps: {
+      ...longPressProps,
+      onContextMenu: callAllHandlers((event: AnyPointerEvent) => {
+        event.preventDefault()
+        onOpen(event)
+        openAndFocusFirstItem()
+      }, props.onContextMenu as any),
+      ref: triggerRef,
+    },
+  }
+}
+
 export interface ContextMenuTriggerProps extends HTMLChakraProps<'span'> {}
 
 export const ContextMenuTrigger: React.FC<ContextMenuTriggerProps> = (
   props
 ) => {
   const { children, ...rest } = props
-  const { triggerRef, onOpen } = useContextMenuContext()
 
-  const menu = useMenuContext()
+  const { triggerProps } = useContextMenuTrigger(props)
 
-  const { openAndFocusFirstItem } = menu
-
-  // @todo add long press support
   return (
     <chakra.span
       {...rest}
       sx={{ WebkitTouchCallout: 'none' }}
-      onContextMenu={(event) => {
-        event.preventDefault()
-        onOpen(event)
-        openAndFocusFirstItem()
-      }}
-      ref={triggerRef}
+      {...triggerProps}
     >
       {children}
     </chakra.span>
@@ -127,21 +217,12 @@ export interface ContextMenuListProps extends MenuListProps {}
 
 export const ContextMenuList: React.FC<ContextMenuListProps> = (props) => {
   const { children, ...rest } = props
-  const { position } = useContextMenuContext()
+  const { menuRef } = useContextMenuContext()
 
   return (
-    <Portal>
-      <MenuList
-        {...rest}
-        style={{
-          position: 'absolute',
-          left: position[0],
-          top: position[1],
-        }}
-      >
-        {children}
-      </MenuList>
-    </Portal>
+    <MenuList ref={menuRef} {...rest}>
+      {children}
+    </MenuList>
   )
 }