+++ apps/website/CHANGELOG.md
@@ -1,5 +1,45 @@
 # website
 
+## 2.3.13
+
+### Patch Changes
+
+- Updated dependencies [bbd692d4]
+- Updated dependencies [781a1c26]
+  - @saas-ui/react@2.5.3
+  - @saas-ui/forms@2.3.7
+
+## 2.3.12
+
+### Patch Changes
+
+- Updated dependencies [042c15f3]
+  - @saas-ui/react@2.5.2
+
+## 2.3.11
+
+### Patch Changes
+
+- @saas-ui/react@2.5.1
+- @saas-ui/forms@2.3.6
+
+## 2.3.10
+
+### Patch Changes
+
+- Updated dependencies [8782a76b]
+- Updated dependencies [8782a76b]
+- Updated dependencies [8782a76b]
+- Updated dependencies [8782a76b]
+- Updated dependencies [8782a76b]
+- Updated dependencies [8782a76b]
+- Updated dependencies [78bc41d0]
+- Updated dependencies [8782a76b]
+- Updated dependencies [316388b9]
+- Updated dependencies [8782a76b]
+  - @saas-ui/react@2.5.0
+  - @saas-ui/forms@2.3.5
+
 ## 2.3.9
 
 ### Patch Changes