+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,42 @@
 # saas-ui-storybook
 
+## 1.1.15
+
+### Patch Changes
+
+- Updated dependencies [bbd692d4]
+- Updated dependencies [781a1c26]
+  - @saas-ui/react@2.5.3
+
+## 1.1.14
+
+### Patch Changes
+
+- Updated dependencies [042c15f3]
+  - @saas-ui/react@2.5.2
+
+## 1.1.13
+
+### Patch Changes
+
+- @saas-ui/react@2.5.1
+
+## 1.1.12
+
+### Patch Changes
+
+- Updated dependencies [8782a76b]
+- Updated dependencies [8782a76b]
+- Updated dependencies [8782a76b]
+- Updated dependencies [8782a76b]
+- Updated dependencies [8782a76b]
+- Updated dependencies [8782a76b]
+- Updated dependencies [78bc41d0]
+- Updated dependencies [8782a76b]
+- Updated dependencies [316388b9]
+- Updated dependencies [8782a76b]
+  - @saas-ui/react@2.5.0
+
 ## 1.1.11
 
 ### Patch Changes